+++ src/init-action.ts
@@ -342,6 +342,7 @@ async function run() {
     config = await initConfig({
       languagesInput: getOptionalInput("languages"),
       queriesInput: getOptionalInput("queries"),
+      qualityQueriesInput: getOptionalInput("quality-queries"),
       packsInput: getOptionalInput("packs"),
       buildModeInput: getOptionalInput("build-mode"),
       configFile,
@@ -587,8 +588,11 @@ async function run() {
       core.exportVariable(bmnVar, value);
     }
 
-    // Set CODEQL_ENABLE_EXPERIMENTAL_FEATURES for rust
-    if (config.languages.includes(Language.rust)) {
+    // For rust: set CODEQL_ENABLE_EXPERIMENTAL_FEATURES, unless codeql already supports rust without it
+    if (
+      config.languages.includes(Language.rust) &&
+      !(await codeql.resolveLanguages()).rust
+    ) {
       const feat = Feature.RustAnalysis;
       const minVer = featureConfig[feat].minimumVersion as string;
       const envVar = "CODEQL_ENABLE_EXPERIMENTAL_FEATURES";